+++ src/components/SystemIntegrations.vue
@@ -18,23 +18,12 @@ defineProps({
     :routeNames="['settingsChannels']"
     forceRemountEvent="forceRemountIntegrations"
     :modelValue="modelValue"
-    :iframeFallback="true"
-    :iframeRoutes="['settingsChannels']"
-    iframeName="integrations"
-    iframeDontUpdateWhenChangesLanguage
     systemClass="system-integrations__system"
-    iframeClass="system-integrations__iframe"
   />
 </template>
 
 <style lang="scss">
 .system-integrations__system {
-  height: 100%;
-  width: 100%;
-}
-
-.system-integrations__iframe {
-  flex: 1;
   overflow: auto;
 }
 </style>